Switch back to source code logic for Proposition.
