+++ src/main/java/de/mpii/clausie/NewProposition.java
@@ -1,9 +1,8 @@
 package de.mpii.clausie;
 
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
+import edu.stanford.nlp.ling.IndexedWord;
+
+import java.util.*;
 
 /**
  * Stores a proposition.
@@ -12,7 +11,7 @@ import java.util.Set;
  *
  * @version $LastChangedRevision: 741 $
  */
-public class OldProposition {
+public class NewProposition {
 
     /**
      * Constituents of the proposition.
@@ -22,11 +21,17 @@ public class OldProposition {
      * Position of optional constituents.
      */
     Set<Integer> optional = new HashSet<>();
+    //IndexedWords
+    private Set<IndexedWord> subjectWords;
+    private Set<IndexedWord> verbWords;
+    private Map<String, Set<IndexedWord>> items;
+    private Object subject;
+    private String type;
 
     // TODO: types of constituents (e.g., optionality)
     // sentence ID etc.
 
-    public OldProposition() {
+    public NewProposition() {
     }
 
     /**
@@ -84,10 +89,46 @@ public class OldProposition {
     }
 
     @Override
-    public OldProposition clone() {
-        OldProposition clone = new OldProposition();
+    public NewProposition clone() {
+        NewProposition clone = new NewProposition();
         clone.constituents = new ArrayList<>(constituents);
         clone.optional = new HashSet<>(optional);
         return clone;
+    }
+
+    public void addSubject(Set<IndexedWord> subjWords) {
+        this.subjectWords = subjWords;
+    }
+
+    public void addVerb(Set<IndexedWord> verbWords) {
+        this.verbWords = verbWords;
+    }
+
+    public void addItem(String type, Set<IndexedWord> words) {
+        if (items == null) {
+            items = new HashMap<>();
+        }
+        items.put(type, words);
+    }
+
+    public Set<IndexedWord> getSubject() {
+        return subjectWords;
+    }
+
+    public Set<IndexedWord> getVerb() {
+        return verbWords;
+    }
+
+    public Map<String, Set<IndexedWord>> getItems() {
+        if (items == null) return Collections.emptyMap();
+        return items;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getType() {
+        return type;
     }
 }
