+++ src/main/java/de/mpii/clausie/NewPropositionGenerator.java
@@ -1,20 +1,18 @@
 package de.mpii.clausie;
 
+import edu.stanford.nlp.ling.CoreLabel;
 import edu.stanford.nlp.ling.IndexedWord;
+import edu.stanford.nlp.ling.Label;
 import edu.stanford.nlp.semgraph.SemanticGraph;
 import edu.stanford.nlp.trees.EnglishGrammaticalRelations;
 import edu.stanford.nlp.trees.GrammaticalRelation;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-import java.util.TreeSet;
+
+import java.util.*;
 
 /**
  * Handles the generation of propositions out of a given clause.
  */
-public abstract class OldPropositionGenerator {
+public abstract class NewPropositionGenerator {
 
     protected final ClausIE clausIE;
     private Set<IndexedWord> words;
@@ -43,14 +41,14 @@ public abstract class OldPropositionGenerator {
     /**
      * Constructs a proposition generator.
      */
-    protected OldPropositionGenerator(ClausIE clausIE) {
+    protected NewPropositionGenerator(ClausIE clausIE) {
         this.clausIE = clausIE;
     }
 
     /**
      * Generates propositions for a given clause.
      */
-    public abstract void generate(List<OldProposition> result, Clause clause, List<Boolean> include);
+    public abstract void generate(List<NewProposition> result, Clause clause, List<Boolean> include);
 
     /**
      * Generates a textual representation of a given constituent plus a set of words.
@@ -106,7 +104,13 @@ public abstract class OldPropositionGenerator {
                            Collection<GrammaticalRelation> excludeRelationsTop) {
         Constituent constituent = clause.constituents.get(constituentIndex);
         if (constituent instanceof TextConstituent) {
-            return ((TextConstituent) constituent).text();
+            String s = ((TextConstituent) constituent).text();
+            words = new TreeSet<>();
+            String[] keys = "value,word,lemma,tag".split(",");
+            Label label = new CoreLabel(keys, createValues(s));
+            IndexedWord word = new IndexedWord(label);
+            words.add(word);
+            return s;
         } else if (constituent instanceof IndexedConstituent) {
             IndexedConstituent iconstituent = (IndexedConstituent) constituent;
             SemanticGraph subgraph = iconstituent.createReducedSemanticGraph();
@@ -126,5 +130,15 @@ public abstract class OldPropositionGenerator {
         } else {
             throw new IllegalArgumentException();
         }
+    }
+
+    private String[] createValues(String s) {
+        if ("has".equals(s)) return "has,has,have,VBZ".split(",");
+        if ("is".equals(s)) return "is,is,be,VBZ".split(",");
+        return (s + ',' + s + ',' + s + ",NN").split(",");
+    }
+
+    public Set<IndexedWord> getWords() {
+        return words;
     }
 }
