+++ src/main/java/de/mpii/clausie/Proposition.java
@@ -1,8 +1,9 @@
 package de.mpii.clausie;
 
-import edu.stanford.nlp.ling.IndexedWord;
-
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 
 /**
  * Stores a proposition.
@@ -21,12 +22,6 @@ public class Proposition {
      * Position of optional constituents.
      */
     Set<Integer> optional = new HashSet<>();
-    //IndexedWords
-    private Set<IndexedWord> subjectWords;
-    private Set<IndexedWord> verbWords;
-    private Map<String, Set<IndexedWord>> items;
-    private Object subject;
-    private String type;
 
     // TODO: types of constituents (e.g., optionality)
     // sentence ID etc.
@@ -94,41 +89,5 @@ public class Proposition {
         clone.constituents = new ArrayList<>(constituents);
         clone.optional = new HashSet<>(optional);
         return clone;
-    }
-
-    public void addSubject(Set<IndexedWord> subjWords) {
-        this.subjectWords = subjWords;
-    }
-
-    public void addVerb(Set<IndexedWord> verbWords) {
-        this.verbWords = verbWords;
-    }
-
-    public void addItem(String type, Set<IndexedWord> words) {
-        if (items == null) {
-            items = new HashMap<>();
-        }
-        items.put(type, words);
-    }
-
-    public Set<IndexedWord> getSubject() {
-        return subjectWords;
-    }
-
-    public Set<IndexedWord> getVerb() {
-        return verbWords;
-    }
-
-    public Map<String, Set<IndexedWord>> getItems() {
-        if (items == null) return Collections.emptyMap();
-        return items;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getType() {
-        return type;
     }
 }
