+++ src/test/java/de/mpii/clausie/TestClausIE.java
@@ -3,6 +3,7 @@ package de.mpii.clausie;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
+import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -20,14 +21,16 @@ public class TestClausIE {
     @BeforeClass
     public static void setUp() {
         inputFile = TestClausIE.class.getResource("task2_ref1.txt");
-        outDir = new File("test_output");
-        outDir.mkdirs();
     }
 
     @Test
-    public void main() throws IOException {
+    public void smokeTest() throws IOException {
         ClausIE.main(new String[]{"-vf",
                                   inputFile.getFile(), "-o",
-                                  outDir + "test"});
+                                  "test.txt"});
+
+        File output = new File("test.txt");
+        Assert.assertTrue(output.exists());
+        output.delete();
     }
 }
